mv scaling params to the corresponding header
